Proyecto/v3/requirements.txt
Flask>=2.2
Flask-Login>=0.6
psycopg2-binary
gunicorn
numpy
pandas
scikit-learn